Return the ResponseLocation before falling back to Location
ResponseLocation [Optional]

    Optionally specifies a different location to which response messages sent as part of
    the protocol or profile should be sent. The allowable syntax of this URI depends on
    the protocol binding.

The ResponseLocation attribute is used to enable different endpoints to be specified for
receiving request and response messages associated with a protocol or profile, not as a
means of load-balancing or redundancy (multiple elements of this type can be included
for this purpose). When a role contains an element of this type pertaining to a protocol
or profile for which only a single type of message (request or response) is applicable,
then the ResponseLocation attribute is unused. [E41]If the ResponseLocation attribute is
omitted, any response messages associated with a protocol or profile may be assumed to
be handled at the URI indicated by the Location attribute.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>